Fixed problem with buy ticket
